requirements.txt
Django==5.0.3
celery==5.3.6
redis==5.0.3
requests==2.31.0
pysocks==1.7.1
APScheduler==3.10.4
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-celery-beat==2.6.0
django-import-export==3.3.7
gunicorn==21.2.0
